requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
networkx>=3.1

# NLP / Sentiment
transformers>=4.35.0
tokenizers>=0.14.0

# RL
gymnasium>=0.29.0

# Data / IO
requests>=2.31.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
einops>=0.7.0

# Viz
matplotlib>=3.7.0
plotly>=5.17.0
rich>=13.0.0

# MLOps
mlflow>=2.7.0

# Dev
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
